extract move_accept_tokens_to_target_kvcache into spec_utils
